Done feed fragment with DB

+++ app/src/main/java/ru/project/iidea/ResponseViewFragmentInterface.java
@@ -2,5 +2,5 @@ package ru.project.iidea;
 
 public interface ResponseViewFragmentInterface {
     public void onBackButtonPressed();
-    public void onUserIdClicked(int userID);
+    public void onUserIdClicked(long userID);
 }
